internal/services/server/modules/exchanger/public_handlers_test.go
package exchanger_test

import (
	"fmt"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/gefion-tech/tg-exchanger-server/internal/services/server"
	"github.com/stretchr/testify/assert"
)

func Test_Server_GetExchangerByNameHandler(t *testing.T) {
	s, redis, teardown := server.TestServer(t)
	defer teardown(redis)

	// Регистрирую менеджера в админке
	tokens, err := server.TestManager(t, s)
	assert.NotNil(t, tokens)
	assert.NoError(t, err)

	assert.NoError(t, server.TestExchanger(t, s, tokens))

	testCases := []struct {
		name         string
		ex_name      string
		expectedCode int
	}{
		{
			name:         "undefined name",
			ex_name:      "undefined",
			expectedCode: http.StatusNotFound,
		},
		{
			name:         "valid",
			ex_name:      "1obmen",
			expectedCode: http.StatusOK,
		},
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			rec := httptest.NewRecorder()
			req, _ := http.NewRequest(http.MethodGet, "/api/v1/admin/exchanger/"+tc.ex_name, nil)
			req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", tokens["access_token"]))
			s.Router.ServeHTTP(rec, req)

			assert.Equal(t, tc.expectedCode, rec.Code)
		})
	}
}
